[mypyc] Refactor: extract callable class generation to new module
